+++ packages/flutter_distributor/lib/src/flutter_distributor.dart
@@ -207,11 +207,38 @@ class FlutterDistributor {
         ProgressBar progressBar = ProgressBar(
           format: 'Publishing to $target: {bar} {value}/{total} {percentage}%',
         );
+
+        Map<String, dynamic>? newPublishArguments = {};
+
+        if (publishArguments != null) {
+          for (var key in publishArguments.keys) {
+            if (!key.startsWith('$target-')) continue;
+            dynamic value = publishArguments[key];
+
+            if (value is List) {
+              value = Map.fromIterable(
+                value,
+                key: (e) => e.split('=')[0],
+                value: (e) => e.split('=')[1],
+              );
+            }
+
+            newPublishArguments.putIfAbsent(
+              key.replaceAll('$target-', ''),
+              () => value,
+            );
+          }
+        }
+
+        if (newPublishArguments.keys.isEmpty) {
+          newPublishArguments = publishArguments;
+        }
+
         PublishResult publishResult = await _publisher.publish(
           file,
           target: target,
           environment: this.environment,
-          publishArguments: publishArguments,
+          publishArguments: newPublishArguments,
           onPublishProgress: (sent, total) {
             if (!progressBar.isActive) {
               progressBar.start(total, sent);
@@ -236,6 +263,8 @@ class FlutterDistributor {
 
   Future<void> release(
     String name, {
+    required List<String> jobNameList,
+    required List<String> skipJobNameList,
     required bool cleanOnceBeforeBuild,
   }) async {
     Directory outputDirectory = distributeOptions.outputDirectory;
@@ -255,7 +284,19 @@ class FlutterDistributor {
     }
 
     for (Release release in releases) {
-      for (ReleaseJob job in release.jobs) {
+      List<ReleaseJob> filteredJobs = release.jobs.where((e) {
+        if (jobNameList.isNotEmpty) {
+          return jobNameList.contains(e.name);
+        }
+        if (skipJobNameList.isNotEmpty) {
+          return !skipJobNameList.contains(e.name);
+        }
+        return true;
+      }).toList();
+      if (filteredJobs.isEmpty) {
+        throw Exception('No available jobs found in ${release.name}.');
+      }
+      for (ReleaseJob job in filteredJobs) {
         List<MakeResult> makeResultList = await package(
           job.package.platform,
           [job.package.target],