+++ packages/flutter_app_packager/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.4
+
+- #22 zip maker supports web platform.
+
 ## 0.1.0
 
 - First release.