+++ packages/flutter_app_publisher/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.4
+
+- Add qiniu publisher.
+
 ## 0.1.0
 
 - First release.