+++ packages/flutter_distributor/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.4
+
+- Add `--version` flag.
+- Add qiniu publisher.
+- #21 Add `--jobs` and `--skip-jobs` option to `release` command.
+- #22 zip maker supports web platform.
+
 ## 0.1.3
 
 - #12 Add `--skip-clean` flag to `package` and `release` commands.